pyproject.toml
[tool.poetry]
name = "test_blog"
version = "0.1.0"
description = ""
authors = ["zaharyak_m <example@email.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.2"
celery = "^4.3"
django-ckeditor = "^5.7"
django-sendgrid-v5 = "^0.8.0"
django-tempus-dominus = "^5.1"
Pillow = "^6.2"
python-dotenv = "^0.10.3"
pytz = "^2019.3"
redis = "^3.3"
sendgrid = "^6.1"
sqlparse = "^0.3.0"
Unidecode = "^1.1"
vine = "^1.3"
zipp = "^0.6.0"
amqp = "^2.5"
billiard = "^3.6"
django-js-asset = "^1.2"
future = "^0.18.0"
importlib-metadata = "^0.23.0"
kombu = "^4.6"
more-itertools = "^7.2"
python-http-client = "^3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
